Separate out ReportBuilder from rpc_recorder.cpp.
