+++ c++11/CMakeLists.txt
@@ -73,6 +73,7 @@ set(LIGHTSTEP_SRCS src/utility.cpp
                    src/propagation.cpp
                    src/binary_carrier.cpp
                    src/grpc_transporter.cpp
+                   src/report_builder.cpp
                    src/rpc_recorder.cpp
                    src/lightstep_span_context.cpp
                    src/lightstep_span.cpp
